Reorganização de lógica da página de transactions.
