--- src/components/common/Drawer/transaction/index.tsx
@@ -1,9 +1,0 @@
-// Imports React
-import React, { Fragment } from 'react'
-
-// Components Imports
-import { GetFormFieldsTransaction } from './formFields'
-
-export const TransactionBody: React.FC = () => {
-  return <Fragment>{GetFormFieldsTransaction()}</Fragment>
-}
