+++ src/components/common/Menu/records/index.tsx
@@ -5,7 +5,7 @@ import React, { useContext } from 'react'
 import { Button, Menu, MenuButton, MenuList } from '@chakra-ui/react'
 
 // Context Imports
-import { ContextDrawer } from '../../../contexts/contextDrawer'
+import { ContextDrawer } from '../../../../contexts/contextDrawer'
 
 // Components Imports
 import MenuItem from '../menuItem'
