--- src/components/pages/transactions/index.ts
@@ -1,1 +1,0 @@
-export * from './Table/index'
