+++ src/contexts/pages/dashboard/index.tsx
@@ -7,8 +7,9 @@ import { ContextProviderProps, DashboardPageContextProps } from '../../types'
 const DashboardPageContext = createContext({} as DashboardPageContextProps)
 
 export const DashboardPageContextProvider = ({ children }: ContextProviderProps) => {
-
   return (
-    <DashboardPageContext.Provider value={{ value: false }}>{children}</DashboardPageContext.Provider>
+    <DashboardPageContext.Provider value={{ value: false }}>
+      {children}
+    </DashboardPageContext.Provider>
   )
 }
