+++ src/contexts/pages/records/index.tsx
@@ -7,7 +7,6 @@ import { ContextProviderProps, RecordsPageContextProps } from '../../types'
 const RecordsPageContext = createContext({} as RecordsPageContextProps)
 
 export const RecordsPageContextProvider = ({ children }: ContextProviderProps) => {
-
   return (
     <RecordsPageContext.Provider value={{ value: false }}>{children}</RecordsPageContext.Provider>
   )
